Main activity: replaced location button with button that leads to list of prefs
PreferencesActivity:
- Minor edits and fixes (added pref repeats when the user opens pref list, fixed)

- Toast message for adding prefs (blank field = error!)

- Fixed PreferenceInfoActivity problem (before: displays the same data. Fixed) + added a few buttons that redirect the user to other activities

+++ app/src/main/java/com/example/metropoliaweatherapp/Preference.java
@@ -4,12 +4,12 @@ import java.util.ArrayList;
 
 public class Preference {
     public String name;
-    public double minTemp;
-    public double maxTemp;
+    public int minTemp;
+    public int maxTemp;
     public String location;
     public String weatherType;
 
-    public Preference(String name, double minTemp, double maxTemp, String location, String weatherType) {
+    public Preference(String name, int minTemp, int maxTemp, String location, String weatherType) {
         ArrayList<Preference> preferences = new ArrayList<>();
 
         this.name = name;
@@ -23,11 +23,11 @@ public class Preference {
         return name;
     }
 
-    public double getMinTemp() {
+    public int getMinTemp() {
         return minTemp;
     }
 
-    public double getMaxTemp() {
+    public int getMaxTemp() {
         return maxTemp;
     }
